+++ lib/models/charity.dart
@@ -6,12 +6,5 @@ class Charity {
   String imagePath;
   String paymentInfo;
 
-  Charity(String name, String id, String description, String website, String imagePath, String paymentInfo) {
-    this.name = name;
-    this.id = id;
-    this.description = description;
-    this.website = website;
-    this.imagePath = imagePath;
-    this.paymentInfo = paymentInfo;
-  }
+  Charity(this.name, this.id, this.description, this.website, this.imagePath, this.paymentInfo);
 }