+++ lib/models/player.dart
@@ -7,19 +7,10 @@ class Player {
   String teamName;
   String league;
   int number;
+  String injured;
   Stats stats;
 
-  Player(String id, String name, String age, String position, String team, String teamName, String league, int number, Stats stats) {
-    this.id = id;
-    this.name = name;
-    this.age = age;
-    this.position = position;
-    this.team = team;
-    this.teamName = teamName;
-    this.league = league;
-    this.number = number;
-    this.stats = stats;
-  }
+  Player(this.id, this.name, this.age, this.position, this.team, this.teamName, this.league, this.number, this.injured, this.stats);
 }
 
 class Stats {
@@ -29,11 +20,5 @@ class Stats {
   int yellowCards;
   int redCards;
 
-  Stats(String appearences, String goals, String assits, String yellowCards, String redCards) {
-    this.appearences = int.parse(appearences);
-    this.goals = int.parse(goals);
-    this.assits = int.parse(assits);
-    this.yellowCards = int.parse(yellowCards);
-    this.redCards = int.parse(redCards);
-  }
+  Stats(this.appearences, this.goals, this.assits, this.yellowCards, this.redCards);
 }