+++ lib/util/themes.dart
@@ -9,4 +9,13 @@ class Themes {
   final textStyle = new TextStyle(
     color: Colors.limeAccent[700],
   );
+
+  Widget playerBarTitle = new Text(
+    'Players',
+    style: new TextStyle(
+      color: Colors.limeAccent[700],
+    ),
+  );
+
+  Icon searchIcon = new Icon(Icons.search, color: Colors.limeAccent[700],);
 }